==> scripts/02 Pre-processing.ipynb <==
!pip install -U ckip-transformers #for tokenization in Traditional Chinese
!pip install TCSP #for stopwords list in Traditional Chinese

import pandas as pd
import re
from ckip_transformers.nlp import CkipWordSegmenter
from TCSP import read_stopwords_list

# read data
df = pd.read_csv('https://raw.githubusercontent.com/Letieateat/Sentiment-Analysis-on-Traditional-Chinese-Lyrics/main/dataset/TW_lyrics.csv')
df = df[df['Lyrics'] != 'Lyrics not found']

# function to clean text
def clean_lyrics(lyric):
   if pd.isna(lyric):
        return ""
   lyric = re.sub(r'[^\u4e00-\u9fff]', '', lyric)
   lyric = re.sub(r'\*+ This Lyrics is NOT for Commercial use \*+', '', lyric)
   lyric = lyric.replace("\n", "")
   return lyric

# clean the text
df['untokenized_lyrics'] = df['Lyrics'].apply(clean_lyrics)

# delete rows without lyrics after cleaning
df = df[df['untokenized_lyrics'] != '']

# initialize drivers
ws_driver  = CkipWordSegmenter(model="bert-base")

# run pipeline for tokenization
df['tokenized_lyrics']  = ws_driver(df['untokenized_lyrics'])

# remove stopwords
stopwords = read_stopwords_list()
df['final_lyrics'] = df['tokenized_lyrics'].apply(lambda x: ' '.join([word for word in x if word not in stopwords]))

# save the cleaned data to a new CSV file
df.to_csv('final_cleaned.csv', index=False)

df=pd.read_csv('https://raw.githubusercontent.com/Letieateat/Sentiment-Analysis-on-Traditional-Chinese-Lyrics/main/dataset/final_cleaned.csv')

# Calculate the total numbers of tokens
total_tokens = df['final_lyrics'].apply(len).sum()

print(f'Total number of tokens in the final_lyrics column: {total_tokens}')
