==> scripts/03 Sentiment Labels.ipynb <==
!pip install pandas

import pandas as pd

# read the dataset
df = pd.read_csv('https://raw.githubusercontent.com/Letieateat/Sentiment-Analysis-on-Traditional-Chinese-Lyrics/main/dataset/final_cleaned.csv')

# defind a function to convert original values into 4 sentiments
def sentiments_4(sentiment):
    if sentiment == 'Angry' or sentiment == 'Anxious':
        return 'Angry/Anxious'
    elif sentiment == 'Excited' or sentiment == 'Happy':
        return 'Excited/Happy'
    elif sentiment == 'Serene' or sentiment == 'Relaxed':
        return 'Serene/Relaxed'
    else:
        return 'Sad'

# defind a function to convert original values into pos/neg sentiments
def sentiments_2(sentiment):
    if sentiment == 'Angry' or sentiment == 'Anxious' or sentiment == 'Sad':
        return 'Negative'
    else:
        return 'Positive'

# apply to the dataset
df['human_4'] = df['human_all'].apply(sentiments_4)
df['human_2'] = df['human_all'].apply(sentiments_2)
print(df.head())

# save the updated dataset
df.to_csv('final_sentiments.csv', index=False)
